+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3896,7 +3896,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M0WHQNSS0XYHGYT8GW2X67NV"),
+  id: z.string().default("01M0WP27XSZNYTBZ730VZMMX06"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),